feat: Refine YNAB MCP
